+++ packages/tokens/src/data/component/circular-progress.ts
@@ -1,0 +1,9 @@
+export const CIRCULAR_PROGRESS = [
+  "containerSize",
+  "activeIndicatorColor",
+  "activeIndicatorThickness",
+  "activeIndicatorShape",
+  "trackColor",
+  "trackThickness",
+  "trackShape",
+] as const;