+++ packages/tokens/src/data/component/index.ts
@@ -1,9 +1,11 @@
+import { CIRCULAR_PROGRESS } from "./circular-progress";
 import { ICON } from "./icon";
 import { MATERIAL_SYMBOL } from "./material-symbols";
 import { RADIO } from "./radio";
 import { RIPPLE } from "./ripple";
 
 export const COMPONENT = {
+  circularProgress: CIRCULAR_PROGRESS,
   icon: ICON,
   radio: RADIO,
   ripple: RIPPLE,